requirements.txt
Django==2.2.4
django-widget-tweaks==1.4.3
numpy==1.16.1
pandas==0.24.1
djangorestframework==3.9.4
openpyxl==2.6.2
django-cors-headers==3.1.0